move handling of ISO durations to helper modules to use in engine and MS
