+++ src/management-system/src/frontend/components/processes/editor/DurationForm.vue
@@ -75,7 +75,7 @@
 </template>
 
 <script>
-import { parseISODuration } from '@/frontend/helpers/instance-information.js';
+import { parseISODuration } from '@proceed/bpmn-helper';
 export default {
   name: 'DurationForm',
   props: {
