+++ src/management-system/src/frontend/helpers/instance-information.js
@@ -69,49 +69,3 @@ export function getPlannedEnd(plannedStart, plannedDuration) {
   anchor.setSeconds(anchor.getSeconds() + (match.groups['ts'] || 0) * direction);
   return anchor;
 }
-
-export function parseISODuration(isoDuration) {
-  let years = null;
-  let months = null;
-  let days = null;
-  let hours = null;
-  let minutes = null;
-  let seconds = null;
-  let dateStr = isoDuration.substring(isoDuration.lastIndexOf('P') + 1);
-  let timeStr = '';
-
-  if (dateStr.includes('T')) {
-    dateStr = isoDuration.substring(isoDuration.lastIndexOf('P') + 1, isoDuration.lastIndexOf('T'));
-    timeStr = isoDuration.substring(isoDuration.lastIndexOf('T') + 1);
-  }
-  if (dateStr.includes('Y')) {
-    const yearsStr = dateStr.substring(0, dateStr.lastIndexOf('Y'));
-    years = parseInt(yearsStr) || null;
-    dateStr = dateStr.substring(dateStr.lastIndexOf('Y') + 1);
-  }
-  if (dateStr.includes('M')) {
-    const monthsStr = dateStr.substring(0, dateStr.lastIndexOf('M'));
-    months = parseInt(monthsStr) || null;
-    dateStr = dateStr.substring(dateStr.lastIndexOf('M') + 1);
-  }
-  if (dateStr.includes('D')) {
-    const daysStr = dateStr.substring(0, dateStr.lastIndexOf('D'));
-    days = parseInt(daysStr) || null;
-  }
-  if (timeStr.includes('H')) {
-    const hoursStr = timeStr.substring(0, timeStr.lastIndexOf('H'));
-    hours = parseInt(hoursStr) || null;
-    timeStr = timeStr.substring(timeStr.lastIndexOf('H') + 1);
-  }
-  if (timeStr.includes('M')) {
-    const minutesStr = timeStr.substring(0, timeStr.lastIndexOf('M'));
-    minutes = parseInt(minutesStr) || null;
-    timeStr = timeStr.substring(timeStr.lastIndexOf('M') + 1);
-  }
-  if (timeStr.includes('S')) {
-    const secondsStr = timeStr.substring(0, timeStr.lastIndexOf('S'));
-    seconds = parseInt(secondsStr) || null;
-  }
-
-  return { years, months, days, hours, minutes, seconds };
-}
